# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gamics API
  version: 0.9.0
  description: >-
    Gamics player, competition, blind match result verification, Gamics result review, rankings, M-Pesa payments,
    notification, identity and role-based operations API.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      tags:
        - Authentication
      summary: Rotate a refresh token and issue a new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Rotated authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    AuthSession:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresInSeconds
        - player
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresInSeconds:
          type: integer
        player:
          $ref: '#/components/schemas/Player'
    Player:
      type: object
      required:
        - id
        - username
        - konamiId
        - displayName
        - email
        - emailVerified
        - phoneNumber
        - countryCode
        - status
      properties:
        id:
          type: string
          format: uuid
        username:
          type:
            - string
            - 'null'
          description: The player's handle, chosen at registration.
        konamiId:
          type:
            - string
            - 'null'
          description: The eFootball Konami ID the player signs in with, as they typed it.
        displayName:
          type: string
          description: The name opponents, match rooms and public profiles show. It starts as the username.
        email:
          type:
            - string
            - 'null'
          format: email
          description: Null until the player adds and verifies one with POST /v1/me/email.
        emailVerified:
          type: boolean
        phoneNumber:
          type:
            - string
            - 'null'
          description: E.164, null until saved with PUT /v1/me/phone.
        countryCode:
          type: string
        birthDate:
          type:
            - string
            - 'null'
          format: date
          description: Asked only when a competition has a minimum age.
        status:
          type: string
        profile:
          oneOf:
            - $ref: '#/components/schemas/Profile'
            - type: 'null'
...
    ProfileInput:
      type: object
      required:
        - handle
        - bio
        - discoverable
        - analyticsConsent
        - scoutingConsent
      properties:
        handle:
          type: string
          pattern: ^[A-Za-z0-9][A-Za-z0-9_.]{2,23}$
        bio:
          type: string
          maxLength: 280
        discoverable:
          type: boolean
        analyticsConsent:
          type: boolean
        scoutingConsent:
          type: boolean
    Profile:
      allOf:
        - $ref: '#/components/schemas/ProfileInput'
...